added input valididation, tweaks to fix putout
changed input text to input number on inputs that are number only
removed unneccicary space in output genration. no harm was done by it but it was unneeded
added css on number fields to remove scroll buttons on any fields set to not scroll
added steamid validation for both input and editing.
swapped default dino settings when loading a file and a creature is missing
added data validation to all number inputs